refactor: makes a clap_app mod and qualifies external crate types
